move webhook route out of /.well-known
